FINALLY! I came back!

+++ structs/metadata.go
@@ -4,12 +4,12 @@ import "time"
 
 type ObjectMetadata struct {
 	ObjectKey     string    `xml:"ObjectKey"`
-	ContentLength int       `xml:"ContentLength"`
+	ContentLength int64     `xml:"ContentLength"`
 	ContentType   string    `xml:"ContentType"`
 	LastModified  time.Time `xml:"LastModified"`
 }
 
-func NewObjectMetadata(ObjectKey string, ContentLength int, ContentType string, LastModified time.Time) *ObjectMetadata {
+func NewObjectMetadata(ObjectKey string, ContentLength int64, ContentType string, LastModified time.Time) *ObjectMetadata {
 	return &ObjectMetadata{
 		ObjectKey:     ObjectKey,
 		ContentLength: ContentLength,
